backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
faiss-cpu==1.8.0
fastapi==0.115.0
frozenlist==1.8.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.3.0
langchain-groq==1.1.2
langchain-text-splitters==0.2.4
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
msgpack==1.1.2
multidict==6.7.1
numpy==1.26.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
propcache==0.4.1
psycopg2-binary==2.9.9
pydantic==2.8.0
pydantic_core==2.20.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.38.6
tenacity==8.5.0
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.30.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
